rename infoModal
