+++ src/lib/components/InfoModal.svelte
